tamc-backend/requirements.txt
# ==========================================
# TAMC Backend Dependencies
# Cleaned and organized - 2025-11-12
# ==========================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
pymysql>=1.1.0
sqlalchemy>=2.0.0

# Deep Learning & Forecasting
torch>=2.0.0
pytorch-lightning>=2.0.0
pytorch-forecasting>=1.0.0
xgboost>=2.0.0

# HTTP & API
requests>=2.31.0
pydantic>=2.0.0

# AI/ML Tools
openai>=1.3.0
google-generativeai>=0.3.0

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0
textblob>=0.17.0

# Utilities
python-dateutil>=2.8.0
joblib>=1.3.0
threadpoolctl>=3.1.0
